+++ src/normalized-responses/built-in-responses/successes/deleted-response.ts
@@ -1,7 +1,7 @@
 import { HttpStatus } from '@nestjs/common';
-import { ParsedExecContextObject } from '../../structure-objects/parsed-exec-context.object';
-import { NormalizedResponseEntryObject } from '../../structure-objects/normalized-response.object';
-import NormalizedResponse from '../normalized-response';
+import { ParsedExecContextObject } from '../../../structure-objects/parsed-exec-context.object';
+import { NormalizedResponseEntryObject } from '../../../structure-objects/normalized-response.object';
+import NormalizedResponse from '../../normalized-response';
 
 export default class DeletedResponse extends NormalizedResponse {
   constructor(
